refactor(internal): extract IntLessonDetail from IntLessonDrawer
The drawer's body (eyebrow, title, summary, CTA, metadata dl) is pure
presentation and will be reused by the split-view detail rail in the
next step. Extract it into IntLessonDetail so both the drawer and the
split-view container can render the same markup.

IntLessonDrawer now only owns the Dialog/Transition chrome and the
close button; IntLessonDetail owns the content. No UI change.

Part of Phase 2 of the internal design system.
