+++ src/components/Internal/IntLessonDrawer.tsx
@@ -1,58 +1,15 @@
 import { Fragment } from 'react';
 import { Dialog, DialogPanel, Transition, TransitionChild } from '@headlessui/react';
-import { ExternalLink, X } from 'lucide-react';
+import { X } from 'lucide-react';
 import type { Lesson } from '@/types';
-import { FILTER_CONFIGS } from '@/utils/filterDefinitions';
-import { intActivityLabel, intGradesLabel } from './IntListRow';
+import { IntLessonDetail } from './IntLessonDetail';
 
 interface IntLessonDrawerProps {
   lesson: Lesson | null;
   onClose: () => void;
 }
 
-function culturalLabel(value: string): string {
-  for (const region of FILTER_CONFIGS.culturalHeritage.options) {
-    if (region.value === value) return region.label;
-    for (const child of region.children ?? []) {
-      if (child.value === value) return child.label;
-    }
-  }
-  return value;
-}
-
-function academicSelected(ai: Lesson['metadata']['academicIntegration']): string[] {
-  if (!ai) return [];
-  if (Array.isArray(ai)) return ai;
-  return ai.selected ?? [];
-}
-
-interface MetaRowProps {
-  label: string;
-  items: string[];
-
-  format?: (value: string) => string;
-}
-
-function MetaRow({ label, items, format }: MetaRowProps) {
-  if (!items || items.length === 0) return null;
-  return (
-    <div className="int-detail-meta-item">
-      <dt>{label}</dt>
-      <dd>
-        {items.map((item) => (
-          <span key={item} className="int-detail-tag">
-            {format ? format(item) : item}
-          </span>
-        ))}
-      </dd>
-    </div>
-  );
-}
-
 export function IntLessonDrawer({ lesson, onClose }: IntLessonDrawerProps) {
-  const meta = lesson?.metadata;
-  const activity = lesson ? intActivityLabel(lesson) : null;
-
   return (
     <Transition show={lesson !== null} as={Fragment}>
       <Dialog open={lesson !== null} onClose={onClose} className="relative z-50">
@@ -79,7 +36,7 @@ export function IntLessonDrawer({ lesson, onClose }: IntLessonDrawerProps) {
             leaveTo="translate-x-full"
           >
             <DialogPanel className="int-drawer pointer-events-auto">
-              {lesson && meta && activity && (
+              {lesson && (
                 <>
                   <button
                     type="button"
@@ -89,41 +46,7 @@ export function IntLessonDrawer({ lesson, onClose }: IntLessonDrawerProps) {
                   >
                     <X width={16} height={16} />
                   </button>
-                  <div className="int-detail-eyebrow">
-                    {activity.label} · Grades {intGradesLabel(lesson.gradeLevels)}
-                  </div>
-                  <h2 className="int-detail-title">{lesson.title}</h2>
-                  <p className="int-detail-summary">{lesson.summary}</p>
-                  {lesson.fileLink && (
-                    <a
-                      className="int-detail-cta"
-                      href={lesson.fileLink}
-                      target="_blank"
-                      rel="noopener noreferrer"
-                    >
-                      Open Lesson Plan
-                      <ExternalLink width={11} height={11} />
-                    </a>
-                  )}
-                  <dl className="int-detail-meta-list">
-                    <MetaRow label="Grades" items={lesson.gradeLevels} />
-                    <MetaRow label="Location" items={meta.locationRequirements ?? []} />
-                    <MetaRow label="Season" items={meta.seasonTiming ?? []} />
-                    <MetaRow label="Themes" items={meta.thematicCategories ?? []} />
-                    <MetaRow label="Competencies" items={meta.coreCompetencies} />
-                    <MetaRow
-                      label="Cultural"
-                      items={meta.culturalHeritage}
-                      format={culturalLabel}
-                    />
-                    <MetaRow label="Academic" items={academicSelected(meta.academicIntegration)} />
-                    <MetaRow label="Garden Skills" items={meta.gardenSkills ?? []} />
-                    <MetaRow label="Cooking Skills" items={meta.cookingSkills ?? []} />
-                    <MetaRow label="Ingredients" items={meta.mainIngredients ?? []} />
-                    <MetaRow label="Cooking Method" items={meta.cookingMethods ?? []} />
-                    <MetaRow label="SEL" items={meta.socialEmotionalLearning ?? []} />
-                    <MetaRow label="Observances" items={meta.observancesHolidays ?? []} />
-                  </dl>
+                  <IntLessonDetail lesson={lesson} />
                 </>
               )}
             </DialogPanel>
