+++ src/components/Internal/index.ts
@@ -2,6 +2,7 @@ export { IntActivePills } from './IntActivePills';
 export { IntCulturalHeritageSection } from './IntCulturalHeritageSection';
 export { IntEmptyState } from './IntEmptyState';
 export { IntFilterSection } from './IntFilterSection';
+export { IntLessonDetail } from './IntLessonDetail';
 export { IntLessonDrawer } from './IntLessonDrawer';
 export { IntListRow, intActivityLabel, intGradesLabel } from './IntListRow';
 export { IntSidebar } from './IntSidebar';
